+++ R/methods_betareg.R
@@ -32,7 +32,7 @@ model_parameters.betareg <- function(model,
 
   ## TODO check merge by
 
-  args <- list(
+  fun_args <- list(
     model,
     ci = ci,
     component = component,
@@ -48,9 +48,9 @@ model_parameters.betareg <- function(model,
     vcov = NULL,
     vcov_args = NULL
   )
-  args <- c(args, dot_args)
+  fun_args <- c(fun_args, dot_args)
 
-  out <- do.call(".model_parameters_generic", args)
+  out <- do.call(".model_parameters_generic", fun_args)
   attr(out, "object_name") <- insight::safe_deparse_symbol(substitute(model))
   out
 }