+++ R/methods_lme4.R
@@ -197,7 +197,7 @@ model_parameters.merMod <- function(model,
         }
       }
     } else {
-      args <- list(
+      fun_args <- list(
         model,
         ci = ci,
         ci_method = ci_method,
@@ -212,8 +212,8 @@ model_parameters.merMod <- function(model,
         vcov = vcov,
         vcov_args = vcov_args
       )
-      args <- c(args, dots)
-      params <- do.call(".extract_parameters_mixed", args)
+      fun_args <- c(fun_args, dots)
+      params <- do.call(".extract_parameters_mixed", fun_args)
     }
 
     params$Effects <- "fixed"
@@ -244,10 +244,10 @@ model_parameters.merMod <- function(model,
     params$Level <- NA
     params$Group <- ""
 
-    if (!is.null(params_random)) {
-      params <- params[match(colnames(params_random), colnames(params))]
-    } else {
+    if (is.null(params_random)) {
       params <- params[match(colnames(params_variance), colnames(params))]
+    } else {
+      params <- params[match(colnames(params_random), colnames(params))]
     }
   }
 
@@ -348,12 +348,12 @@ standard_error.merMod <- function(model,
   }
 
   if (!is.null(vcov) || isTRUE(dots[["robust"]])) {
-    args <- list(model,
+    fun_args <- list(model,
       vcov = vcov,
       vcov_args = vcov_args
     )
-    args <- c(args, dots)
-    out <- do.call("standard_error.default", args)
+    fun_args <- c(fun_args, dots)
+    out <- do.call("standard_error.default", fun_args)
     return(out)
   }